backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
aiofiles==23.2.1
email-validator==2.1.0
python-keycloak==3.8.0
cors==1.0.1
python-dotenv==1.0.0
alembic==1.12.1
tenacity==8.2.3
